GameData: Implement basic scoring
The current scoring set up is a little messy since some information is
maintained by the Players's individual Counters and some being of a global
nature must be determined by the GameData class.  Perhaps it would be best to
move all of this information into GameData or at least keep track of which
players have the most cards and most spades to avoid unnecessary recalculation.